+++ ssdaq/core/data/slowsignal_format.py
