+++ ssdaq/core/data/trigger_format.py
