--- ssdaq/receivers.py
@@ -1,5 +1,0 @@
-from .core.slowsignal.readout_assembler import SSReadoutAssembler
-from .core.timestamps.timestamp_receiver import TimestampReceiver
-from .core.logging.log_receiver import LogReceiver
-from .core.triggers.trigger_receiver import TriggerPacketReceiver
-from .core.monitor.monitor_receiver import MonitorReceiver
