+++ ssdaq/receivers/monitor_receiver.py
@@ -1,10 +1,6 @@
 import asyncio
-from ssdaq.core.receiver_server import ReceiverServer,ReceiverMonSender
-
 import zmq
-
-from ssdaq.core.timestamps import CDTS_pb2
-
+from ssdaq.core.receiver_server import ReceiverServer, ReceiverMonSender
 
 class MonitorReceiver(ReceiverServer):
     def __init__(self, ip: str, port: int, publishers: list):
@@ -18,6 +14,7 @@ class MonitorReceiver(ReceiverServer):
         self.receiver.bind(connectionstr)
         self._setup = True
         self.mon = ReceiverMonSender("MonitorReceiver", self.loop, self._context)
+
     async def ct_subscribe(self):
         while self.running:
             packet = await self.receiver.recv()
