+++ ssdaq/receivers/trigger_receiver.py
